Make BPF symbolization tests end-to-end tests
The BPF symbolization tests were added as unit tests, because they
require a bunch of setup shenanigans that we only had internally to the
crate. Conceptually, however, they are very much end-to-end tests that
test the public API surface and not just some implementation detail. We
*really* want to preserve the ability to get a good overview of all
end-to-end tests in a single location and so with this change we take
necessary steps to expose necessary test helpers so that they can be
used from end-to-end tests. Furthermore, we move said BPF symbolization
tests into tests/blazesym.rs, co-located with most other end-to-end
tests.
