+++ src/symbolize/symbolizer.rs
@@ -1375,17 +1375,12 @@ impl Default for Symbolizer {
 mod tests {
     use super::*;
 
-    #[cfg(all(target_os = "linux", feature = "nightly"))]
-    use test::Bencher;
-
     use test_log::test;
 
     use crate::maps::Perm;
     use crate::symbolize;
     use crate::symbolize::CodeInfo;
     use crate::test_helper::find_the_answer_fn_in_zip;
-    #[cfg(target_os = "linux")]
-    use crate::test_helper::with_bpf_symbolization_target_addrs;
 
 
     /// Exercise the `Debug` representation of various types.
@@ -1691,87 +1686,5 @@ mod tests {
 
         let () = test(zip_error_dispatch);
         let () = test(zip_delayed_error_dispatch);
-    }
-
-    /// Test symbolization of a kernel address inside a BPF program.
-    #[cfg(target_os = "linux")]
-    #[test]
-    fn symbolize_kernel_bpf_program() {
-        with_bpf_symbolization_target_addrs(|handle_getpid, subprogram| {
-            let src = symbolize::Source::Kernel(symbolize::Kernel::default());
-            let symbolizer = Symbolizer::new();
-            let result = symbolizer
-                .symbolize(
-                    &src,
-                    symbolize::Input::AbsAddr(&[handle_getpid, subprogram]),
-                )
-                .unwrap();
-            let handle_getpid_sym = result[0].as_sym().unwrap();
-            assert_eq!(handle_getpid_sym.name, "handle__getpid");
-            let code_info = handle_getpid_sym.code_info.as_ref().unwrap();
-            assert_eq!(code_info.dir, None);
-            assert_eq!(
-                Path::new(&code_info.file).file_name(),
-                Some(OsStr::new("getpid.bpf.c"))
-            );
-            assert_eq!(code_info.line, Some(33));
-            assert_ne!(code_info.column, None);
-
-            let subprogram_sym = result[1].as_sym().unwrap();
-            assert_eq!(subprogram_sym.name, "subprogram");
-            let code_info = subprogram_sym.code_info.as_ref().unwrap();
-            assert_eq!(code_info.dir, None);
-            assert_eq!(
-                Path::new(&code_info.file).file_name(),
-                Some(OsStr::new("getpid.bpf.c"))
-            );
-            assert_eq!(code_info.line, Some(15));
-            assert_ne!(code_info.column, None);
-        })
-    }
-
-    /// Benchmark the symbolization of BPF program kernel addresses.
-    #[cfg(target_os = "linux")]
-    #[cfg(feature = "nightly")]
-    #[bench]
-    fn bench_symbolize_kernel_bpf_uncached(b: &mut Bencher) {
-        with_bpf_symbolization_target_addrs(|handle_getpid, subprogram| {
-            let () = b.iter(|| {
-                let src = symbolize::Source::Kernel(symbolize::Kernel::default());
-                let symbolizer = Symbolizer::new();
-
-                let result = symbolizer
-                    .symbolize(
-                        &src,
-                        symbolize::Input::AbsAddr(&[handle_getpid, subprogram]),
-                    )
-                    .unwrap();
-
-                assert_eq!(result.len(), 2);
-            });
-        });
-    }
-
-    /// Benchmark the symbolization of BPF program kernel addresses when
-    /// relevant data is readily cached.
-    #[cfg(target_os = "linux")]
-    #[cfg(feature = "nightly")]
-    #[bench]
-    fn bench_symbolize_kernel_bpf_cached(b: &mut Bencher) {
-        with_bpf_symbolization_target_addrs(|handle_getpid, subprogram| {
-            let src = symbolize::Source::Kernel(symbolize::Kernel::default());
-            let symbolizer = Symbolizer::new();
-
-            let () = b.iter(|| {
-                let result = symbolizer
-                    .symbolize(
-                        &src,
-                        symbolize::Input::AbsAddr(&[handle_getpid, subprogram]),
-                    )
-                    .unwrap();
-
-                assert_eq!(result.len(), 2);
-            });
-        });
     }
 }
